CROSSLINK-302 on closing action, requester calls DeleteItem (#745)
Only call DeleteItem if AcceptItem succeeds for each item
Store the exact item ID passed to AcceptItem in a dedicated lms_item_id column and
use it for DeleteItem cleanup.

+++ broker/lms/lms_adapter_ncip.go
@@ -2,6 +2,7 @@ package lms
 
 import (
 	"encoding/xml"
+	"errors"
 	"fmt"
 	"net/http"
 	"strings"
@@ -281,10 +282,17 @@ func (l *LmsAdapterNcip) AcceptItem(
 }
 
 func (l *LmsAdapterNcip) DeleteItem(itemId string) error {
+	if l.config.AcceptItemEnabled != nil && !*l.config.AcceptItemEnabled {
+		return nil
+	}
 	arg := ncip.DeleteItem{
 		ItemId: ncip.ItemId{ItemIdentifierValue: itemId},
 	}
 	_, err := l.ncipClient.DeleteItem(arg)
+	var ncipErr *ncipclient.NcipError
+	if errors.As(err, &ncipErr) && ncipErr.Problem.ProblemType.Text == string(ncip.UnknownItem) {
+		return nil
+	}
 	return err
 }
